requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.4
psycopg2-binary>=2.9
dj-database-url>=2.2
django-environ>=0.11
python-dateutil>=2.9
rasterio
requests>=2.32.0

